# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biilche.ir",
        "description": "Biilche Application API",
        "contact": {
            "email": "ansaramman@gmail.com"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://admin.biilche.ir/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/sendQuestions": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Store New Question",
                "description": "Store New Question",
                "operationId": "storeQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreQuestionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/questions/toggleToFavorite": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Add/Delete from to user favorite list",
                "description": "Add/Delete from to user favorite list",
                "operationId": "toggleToFavorite",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreQuestionRequest": {
                "title": "Store Question request",
                "description": "Store Question request body data",
                "required": [
                    "content"
                ],
                "properties": {
                    "content": {
                        "title": "content",
                        "description": "Content of the new Question",
                        "type": "string",
                        "example": "A nice Question"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}